[WEJBHTTP-152] Refactoring - renaming RemoteHTTPHandler to AbstractEjbHandler and making it inner class of ServerHandlers
